requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
numpy>=1.21.0
Pillow>=8.3.0
scipy>=1.7.0
matplotlib>=3.4.0
soundfile>=0.10.0
librosa>=0.8.0
